+++ test/Sentry.Tests/Protocol/TransactionTests.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using FluentAssertions;
 using NSubstitute;
 using Sentry.Extensibility;
@@ -119,10 +120,33 @@ namespace Sentry.Tests.Protocol
         }
 
         [Fact]
+        public void StartChild_Limit_Maintained()
+        {
+            // Arrange
+            var transaction = new TransactionTracer(DisabledHub.Instance, "my name", "my op")
+            {
+                IsSampled = true
+            };
+
+            // Act
+            var spans = Enumerable
+                .Range(0, 1000 * 2)
+                .Select(i => transaction.StartChild("span " + i))
+                .ToArray();
+
+            // Assert
+            transaction.Spans.Should().HaveCount(1000);
+            spans.Count(s => s.IsSampled == true).Should().Be(1000);
+        }
+
+        [Fact]
         public void StartChild_SamplingInherited_Null()
         {
             // Arrange
-            var transaction = new TransactionTracer(DisabledHub.Instance, "my name", "my op") {IsSampled = null};
+            var transaction = new TransactionTracer(DisabledHub.Instance, "my name", "my op")
+            {
+                IsSampled = null
+            };
 
             // Act
             var child = transaction.StartChild("child op", "child desc");
@@ -135,7 +159,10 @@ namespace Sentry.Tests.Protocol
         public void StartChild_SamplingInherited_True()
         {
             // Arrange
-            var transaction = new TransactionTracer(DisabledHub.Instance, "my name", "my op") {IsSampled = true};
+            var transaction = new TransactionTracer(DisabledHub.Instance, "my name", "my op")
+            {
+                IsSampled = true
+            };
 
             // Act
             var child = transaction.StartChild("child op", "child desc");
@@ -148,13 +175,34 @@ namespace Sentry.Tests.Protocol
         public void StartChild_SamplingInherited_False()
         {
             // Arrange
-            var transaction = new TransactionTracer(DisabledHub.Instance, "my name", "my op") {IsSampled = false};
+            var transaction = new TransactionTracer(DisabledHub.Instance, "my name", "my op")
+            {
+                IsSampled = false
+            };
 
             // Act
             var child = transaction.StartChild("child op", "child desc");
 
             // Assert
             child.IsSampled.Should().BeFalse();
+        }
+
+        [Fact]
+        public void StartChild_TraceIdInherited()
+        {
+            // Arrange
+            var transaction = new TransactionTracer(DisabledHub.Instance, "my name", "my op");
+
+            // Act
+            var children = new[]
+            {
+                transaction.StartChild("op1"),
+                transaction.StartChild("op2"),
+                transaction.StartChild("op3")
+            };
+
+            // Assert
+            children.Should().OnlyContain(s => s.TraceId == transaction.TraceId);
         }
 
         [Fact]
@@ -176,7 +224,8 @@ namespace Sentry.Tests.Protocol
         {
             // Arrange
             var client = Substitute.For<ISentryClient>();
-            var hub = new Hub(client, new SentryOptions{Dsn = DsnSamples.ValidDsnWithoutSecret});
+            var options = new SentryOptions {Dsn = DsnSamples.ValidDsnWithoutSecret};
+            var hub = new Hub(client, options);
 
             var transaction = new TransactionTracer(hub, "my name", "my op");
 
@@ -188,11 +237,39 @@ namespace Sentry.Tests.Protocol
         }
 
         [Fact]
-        public void Finish_LinksExceptionToEvent()
+        public void Finish_DropsUnfinishedSpans()
         {
             // Arrange
             var client = Substitute.For<ISentryClient>();
             var hub = new Hub(client, new SentryOptions{Dsn = DsnSamples.ValidDsnWithoutSecret});
+
+            var transaction = new TransactionTracer(hub, "my name", "my op");
+
+            transaction.StartChild("op1").Finish();
+            transaction.StartChild("op2");
+            transaction.StartChild("op3").Finish();
+
+            // Act
+            transaction.Finish();
+
+            // Assert
+            client.Received(1).CaptureTransaction(
+                Arg.Is<Transaction>(t =>
+                    t.Spans.Count == 2 &&
+                    t.Spans.Any(s => s.Operation == "op1") &&
+                    t.Spans.All(s => s.Operation != "op2") &&
+                    t.Spans.Any(s => s.Operation == "op3")
+                )
+            );
+        }
+
+        [Fact]
+        public void Finish_LinksExceptionToEvent()
+        {
+            // Arrange
+            var client = Substitute.For<ISentryClient>();
+            var options = new SentryOptions {Dsn = DsnSamples.ValidDsnWithoutSecret};
+            var hub = new Hub(client, options);
 
             var exception = new InvalidOperationException();
             var transaction = new TransactionTracer(hub, "my name", "my op");